separating secrets provider and secret info into separate files as per review
